+++ addon/globalPlugins/chessmart/virtual_chessboard/__init__.py
@@ -4,11 +4,7 @@
 from .base import BaseVirtualChessboard
 from .user_engine import UserEngineChessboard
 from .user_user import UserUserChessboard
-from .pgn_player import (
-	PGNPlayerChessboard,
-	PGNGame,
-	PGNGameInfo,
-)
+from .pgn_player import PGNPlayerChessboard
 from .puzzle_board import PuzzleChessboard
 from .endgame_board import EndgameDrillChessboard, EndgameLessonChessboard
 
@@ -17,8 +13,6 @@ __all__ = [
 	"UserEngineChessboard",
 	"UserUserChessboard",
 	"PGNPlayerChessboard",
-	"PGNGame",
-	"PGNGameInfo",
 	"PuzzleChessboard",
 	"EndgameDrillChessboard",
 	"EndgameLessonChessboard",